add graphHelpers.js
